% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDRIFT.R
\name{CellDRIFT}
\alias{CellDRIFT}
\title{CellDRIFT}
\usage{
CellDRIFT(DNAm)
}
\arguments{
\item{DNAm}{A matrix or data frame that contains the DNA methylation beta values as obtained from Illumina 450k or EPIC arrays. Columns should be named according to CpG ID but do not have to be in a specific order.}
}
\value{
Will return a vector of predicted CellDrift values with length equal to the number of input rows (or samples).
}
\description{
A function to run the appropriate PC fit and obtain CellDRIFT for samples in a single line of code.
}
\examples{
CellDRIFT(DNAm = exampleDNAm)
}
