% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_tan.R
\docType{data}
\name{PCA_tan}
\alias{PCA_tan}
\title{PCA_tan}
\format{
A list object as defined by the output of calling prcomp([tan-module-CpG-Mat])
\describe{
  \item{sdev}{Standard deviation of principal components--the sqrt of the singular vector eigenevalues.}
  \item{rotation}{The matrix of CpG loadings.}
  \item{center}{The values used to center by CpG in the new data matrix.}
  \item{scale}{Indicates that CpGs were not scaled.}
  \item{x}{The value of the rotated samples on the new PC features.}
}
}
\usage{
PCA_tan
}
\description{
The PCA object allowing prediction and mapping of tan PCs in new data.
}
\keyword{datasets}
