% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tan_module_means.R
\docType{data}
\name{tan_module_means}
\alias{tan_module_means}
\title{tan_module_means}
\format{
An object of class \code{numeric} of length 297.
}
\usage{
tan_module_means
}
\description{
The vector giving the means of the tan module in the original training data which can be used for imputation in the case that a column of CpGs is missing after cleaning.
}
\keyword{datasets}
