% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yellow_module_means.R
\docType{data}
\name{yellow_module_means}
\alias{yellow_module_means}
\title{yellow_module_means}
\format{
An object of class \code{numeric} of length 2025.
}
\usage{
yellow_module_means
}
\description{
The vector giving the means of the yellow module in the original training data which can be used for imputation in the case that a column of CpGs is missing after cleaning.
}
\keyword{datasets}
